move auto min/max out of applyField overrides
