+++ packages/grafana-data/src/field/fieldOverrides.test.ts
@@ -2,7 +2,6 @@ import {
   applyFieldOverrides,
   applyRawFieldOverrides,
   FieldOverrideEnv,
-  findNumericFieldMinMax,
   getLinksSupplier,
   setDynamicConfigValue,
   setFieldConfigDefaults,
@@ -73,68 +72,6 @@ locationUtil.initialize({
   buildParamsFromVariables: () => {},
   // @ts-ignore
   getTimeRangeForUrl: () => {},
-});
-
-describe('Global MinMax', () => {
-  it('find global min max', () => {
-    const f0 = new ArrayDataFrame<{ title: string; value: number; value2: number | null }>([
-      { title: 'AAA', value: 100, value2: 1234 },
-      { title: 'BBB', value: -20, value2: null },
-      { title: 'CCC', value: 200, value2: 1000 },
-    ]);
-
-    const minmax = findNumericFieldMinMax([f0]);
-    expect(minmax.min).toEqual(-20);
-    expect(minmax.max).toEqual(1234);
-  });
-
-  it('find global min max when all values are zero', () => {
-    const f0 = new ArrayDataFrame<{ title: string; value: number; value2: number | null }>([
-      { title: 'AAA', value: 0, value2: 0 },
-      { title: 'CCC', value: 0, value2: 0 },
-    ]);
-
-    const minmax = findNumericFieldMinMax([f0]);
-    expect(minmax.min).toEqual(0);
-    expect(minmax.max).toEqual(0);
-  });
-
-  describe('when value is null', () => {
-    it('then global min max should be null', () => {
-      const frame = toDataFrame({
-        fields: [
-          { name: 'Time', type: FieldType.time, values: [1] },
-          { name: 'Value', type: FieldType.number, values: [null] },
-        ],
-      });
-      const { min, max } = findNumericFieldMinMax([frame]);
-
-      expect(min).toBe(null);
-      expect(max).toBe(null);
-    });
-  });
-
-  describe('when value values are zeo', () => {
-    it('then global min max should be correct', () => {
-      const frame = toDataFrame({
-        fields: [
-          { name: 'Time', type: FieldType.time, values: [1, 2] },
-          { name: 'Value', type: FieldType.number, values: [1, 2] },
-        ],
-      });
-      const frame2 = toDataFrame({
-        fields: [
-          { name: 'Time', type: FieldType.time, values: [1, 2] },
-          { name: 'Value', type: FieldType.number, values: [0, 0] },
-        ],
-      });
-
-      const { min, max } = findNumericFieldMinMax([frame, frame2]);
-
-      expect(min).toBe(0);
-      expect(max).toBe(2);
-    });
-  });
 });
 
 describe('applyFieldOverrides', () => {
@@ -293,7 +230,6 @@ describe('applyFieldOverrides', () => {
       replaceVariables: (undefined as any) as InterpolateFunction,
       getDataSourceSettingsByUid: undefined as any,
       theme: getTestTheme(),
-      autoMinMax: true,
     })[0];
     const valueColumn = data.fields[1];
     const config = valueColumn.config;
@@ -317,7 +253,6 @@ describe('applyFieldOverrides', () => {
       }) as InterpolateFunction,
       getDataSourceSettingsByUid: undefined as any,
       theme: getTestTheme(),
-      autoMinMax: true,
       fieldConfigRegistry: customFieldRegistry,
     })[0];
